# api/contracts/control-plane.openapi.json
{
  "components": {
    "schemas": {
...
      "DomainMetadataSnapshotResponse-Output": {
        "properties": {
          "entries": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DomainMetadataResponse"
            },
            "title": "Entries",
            "type": "object"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "title": "DomainMetadataSnapshotResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegimePolicy": {
        "additionalProperties": false,
        "properties": {
          "blockOnTransition": {
            "default": true,
            "title": "Blockontransition",
            "type": "boolean"
          },
          "blockOnUnclassified": {
            "default": true,
            "title": "Blockonunclassified",
            "type": "boolean"
          },
          "honorHaltFlag": {
            "default": true,
            "title": "Honorhaltflag",
            "type": "boolean"
          },
          "modelName": {
            "default": "default-regime",
            "maxLength": 128,
            "minLength": 1,
            "title": "Modelname",
            "type": "string"
          },
          "onBlocked": {
            "default": "skip_entries",
            "enum": [
              "skip_entries",
              "skip_rebalance"
            ],
            "title": "Onblocked",
            "type": "string"
          },
          "targetGrossExposureByRegime": {
            "$ref": "#/components/schemas/TargetGrossExposureByRegime"
          }
        },
        "title": "RegimePolicy",
        "type": "object"
      },
...
      "StrategyConfig-Input": {
        "additionalProperties": false,
        "properties": {
          "costModel": {
            "default": "default",
            "maxLength": 64,
            "minLength": 1,
            "title": "Costmodel",
            "type": "string"
          },
          "exits": {
            "items": {
              "$ref": "#/components/schemas/ExitRule"
            },
            "title": "Exits",
            "type": "array"
          },
          "holdingPeriod": {
            "default": 21,
            "minimum": 1.0,
            "title": "Holdingperiod",
            "type": "integer"
          },
          "intrabarConflictPolicy": {
            "default": "stop_first",
            "enum": [
              "stop_first",
              "take_profit_first",
              "priority_order"
            ],
            "title": "Intrabarconflictpolicy",
            "type": "string"
          },
          "longOnly": {
            "default": true,
            "title": "Longonly",
            "type": "boolean"
          },
          "lookbackWindow": {
            "default": 63,
            "minimum": 1.0,
            "title": "Lookbackwindow",
            "type": "integer"
          },
          "rankingSchemaName": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rankingschemaname"
          },
          "rebalance": {
            "default": "monthly",
            "maxLength": 64,
            "minLength": 1,
            "title": "Rebalance",
            "type": "string"
          },
          "regimePolicy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegimePolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "topN": {
            "default": 20,
            "minimum": 1.0,
            "title": "Topn",
            "type": "integer"
          },
          "universe": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UniverseDefinition-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "universeConfigName": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Universeconfigname"
          }
        },
        "title": "StrategyConfig",
        "type": "object"
      },
...
      "StrategySummaryResponse": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_table_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Table Name"
          },
          "type": {
            "default": "configured",
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "name"
        ],
        "title": "StrategySummaryResponse",
        "type": "object"
      },
      "StrategyUpsertRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/StrategyConfig-Input"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "type": {
            "default": "configured",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "config"
        ],
        "title": "StrategyUpsertRequest",
        "type": "object"
      },
...
      "SystemStatusViewResponse": {
        "properties": {
          "generatedAt": {
            "title": "Generatedat",
            "type": "string"
          },
          "metadataSnapshot": {
            "$ref": "#/components/schemas/DomainMetadataSnapshotResponse-Output"
          },
          "sources": {
            "$ref": "#/components/schemas/SystemStatusViewSources"
          },
          "systemHealth": {
            "additionalProperties": true,
            "title": "Systemhealth",
            "type": "object"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "generatedAt",
          "sources"
        ],
        "title": "SystemStatusViewResponse",
        "type": "object"
      },
      "SystemStatusViewSources": {
        "properties": {
          "metadataSnapshot": {
            "const": "persisted-snapshot",
            "default": "persisted-snapshot",
            "title": "Metadatasnapshot",
            "type": "string"
          },
          "systemHealth": {
            "enum": [
              "cache",
              "live-refresh"
            ],
            "title": "Systemhealth",
            "type": "string"
          }
        },
        "required": [
          "systemHealth"
        ],
        "title": "SystemStatusViewSources",
        "type": "object"
      },
...
      "UniverseDefinition-Input": {
        "additionalProperties": false,
        "properties": {
          "root": {
            "$ref": "#/components/schemas/UniverseGroup-Input"
          },
          "source": {
            "const": "postgres_gold",
            "default": "postgres_gold",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "root"
        ],
        "title": "UniverseDefinition",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Asset Allocation API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/strategies/": {
      "get": {
        "description": "List all available strategies.",
        "operationId": "list_strategies_api_strategies__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategySummaryResponse"
                  },
                  "title": "Response List Strategies Api Strategies  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Strategies",
        "tags": [
          "Strategies"
        ]
      },
      "post": {
        "description": "Create or update a strategy configuration.\nRequires authentication.",
        "operationId": "save_strategy_api_strategies__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Save Strategy Api Strategies  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Strategy",
        "tags": [
          "Strategies"
        ]
      }
    },
...
    "/api/system/status-view": {
      "get": {
        "operationId": "system_status_view_api_system_status_view_get",
        "parameters": [
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusViewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "System Status View",
        "tags": [
          "System"
        ]
      }
    },
...
  }
}